replace db content with db-new
