+++ packages/grammar-sdk/index.ts
@@ -1,5 +1,5 @@
 export * from "./src/grammar";
-export type { Exercise } from "./src/exercise";
+export { type Exercise, exerciseSchema } from "./src/exercise";
 export * from "./src/context";
 export type { GrammarPoint } from "./src/grammar-point";
-export { AuthorizationError, isAuthorizationError } from "./src/db-new";
+export { AuthorizationError, isAuthorizationError } from "./src/db";
