+++ packages/grammar-sdk/src/grammar.ts
@@ -1,65 +1,32 @@
-import { createSupabaseClientInstance } from "../../../libs/supabase";
-import { Context } from "./context";
+import type { Context } from "./context";
 import { getGrammarPoint, getGrammarPoints } from "./db";
-import {
-  getGrammarPoint as getGrammarPointNew,
-  getGrammarPoints as getGrammarPointsNew,
-} from "./db-new";
 import { GrammarPoint, GrammarPoints } from "./grammar-point";
 
 export const fetchGrammarPoint = async (
   id: string,
   context: Context,
 ): Promise<GrammarPoint | undefined> => {
-  if (Context.isAdmin(context)) {
-    const maybeGrammarPoint = await getGrammarPointNew(+id);
-    return (
-      maybeGrammarPoint &&
-      GrammarPoint.filterVisible(maybeGrammarPoint, context)
-    );
-  }
-  const grammarPoint = await getGrammarPoint(+id);
-  if (grammarPoint && !GrammarPoint.filterVisible(grammarPoint, context)) {
-    return undefined;
-  }
-  return grammarPoint;
+  const maybeGrammarPoint = await getGrammarPoint(+id);
+  return (
+    maybeGrammarPoint && GrammarPoint.filterVisible(maybeGrammarPoint, context)
+  );
 };
 
 export const fetchGrammarPoints = async (
   ids: string[],
   context: Context,
 ): Promise<GrammarPoint[]> => {
-  if (Context.isAdmin(context)) {
-    return GrammarPoints.filterVisible(
-      await getGrammarPointsNew(ids.map((id) => +id)),
-      context,
-    );
-  }
-  const grammarPoints = await getGrammarPoints(ids.map((id) => +id));
-  return GrammarPoints.filterVisible(grammarPoints, context);
+  return GrammarPoints.filterVisible(
+    await getGrammarPoints(ids.map((id) => +id)),
+    context,
+  );
 };
 
 export const fetchAllGrammarPoints = async (
   context: Context,
 ): Promise<GrammarPoint[]> => {
-  const grammarPoints = Context.isAdmin(context)
-    ? await getGrammarPointsNew()
-    : await getGrammarPoints();
+  const grammarPoints = await getGrammarPoints();
   return GrammarPoints.filterVisible(grammarPoints, context);
-};
-
-export const fetchExplanation = async (grammarPointId: string | number) => {
-  const supabase = createSupabaseClientInstance();
-
-  const url = supabase.storage
-    .from("explanations")
-    .getPublicUrl(`SON-${grammarPointId}.html`).data.publicUrl;
-
-  const response = await fetch(url);
-  if (response.ok) {
-    return await response.text();
-  }
-  return undefined;
 };
 
 export {
@@ -67,4 +34,4 @@ export {
   updateGrammarPoint,
   updateGrammarPointsOrder,
   putExercises,
-} from "./db-new";
+} from "./db";
